configure folder for apps and move app authentication
