+++ apps/authentication/templates/emails/register.html
