+++ apps/authentication/tests/test_pytest.py
@@ -1,0 +1,6 @@
+from django.test import TestCase
+
+
+class StartTest(TestCase):
+    def test_initial(self):
+        self.assertEqual(1, 1)
