+++ bra_back/api.py
@@ -1,6 +1,6 @@
 from ninja import NinjaAPI
 
-from authentication.api import auth_router
+from apps.authentication.api import auth_router
 
 api: NinjaAPI = NinjaAPI()
-api.add_router('users', auth_router)
+api.add_router('/users', auth_router)
