+++ bra_back/settings.py
@@ -19,8 +19,8 @@ import sys
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
 
-# PROJECT_ROOT = os.path.dirname(__file__)
-# sys.path.insert(0, os.path.join(PROJECT_ROOT, '../applications'))
+PROJECT_ROOT = os.path.dirname(__file__)
+sys.path.insert(0, os.path.join(PROJECT_ROOT, '../apps'))
 
 # Quick-start development settings - unsuitable for production
 # See https://docs.djangoproject.com/en/4.1/howto/deployment/checklist/
@@ -44,7 +44,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'ninja',
-    'authentication'
+    'authentication',
 ]
 
 MIDDLEWARE = [
@@ -98,12 +98,6 @@ else:
         }
     }
 
-    # db_options = {
-    #     'default': {
-    #         'ENGINE': 'django.db.backends.sqlite3',
-    #         'NAME': BASE_DIR / 'db.sqlite3'
-    #     }
-    # }
 
 DATABASES = db_options
 
@@ -154,8 +148,8 @@ AUTHENTICATION_BACKENDS = ('authentication.backends.CustomBackend',)
 
 # E-mails
 
-REGISTER_TEMPLATE_PATH: str = os.path.join(BASE_DIR, 'authentication/templates/emails/register.html')
-UPDATE_TEMPLATE_PATH: str = os.path.join(BASE_DIR, 'authentication/templates/emails/update.html')
+REGISTER_TEMPLATE_PATH: str = os.path.join(BASE_DIR, 'apps/authentication/templates/emails/register.html')
+UPDATE_TEMPLATE_PATH: str = os.path.join(BASE_DIR, 'apps/authentication/templates/emails/update.html')
 
 EMAIL_BACKEND: str
 if DEBUG:
